Refactor database and car repository queries; update budget extraction to handle float values; add __init__.py files for service modules

+++ app/services/chatbot/entities.py
@@ -10,7 +10,7 @@ def extract_budget(message: str) -> float | None:
     if not match:
         return None
     
-    value = int(match.group(1))
+    value = float(match.group(1))
     unit = match.group(2)
 
     if unit == "million":
